% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{generate_changelog}
\alias{generate_changelog}
\title{Generate a changelog from two data frames.}
\usage{
generate_changelog(
  .x,
  .y,
  id_cols,
  names_to = "question",
  .x_values_to = "raw_value",
  .y_values_to = "transformed_value"
)
}
\arguments{
\item{.x}{data frame x}

\item{.y}{data frame y}

\item{id_cols}{columns common between x and y that uniquely identify observations}

\item{names_to}{name of the column containing the pivoted column names}

\item{.x_values_to}{name of the column containing values from x}

\item{.y_values_to}{name of the column containing values from y}
}
\value{
data frame
}
\description{
Generate a changelog from two data frames.
}
