% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\name{get_question_columns}
\alias{get_question_columns}
\title{Get the columns corresponding to a \verb{select multiple} question}
\usage{
get_question_columns(.df, question_label, group_sep = "/")
}
\arguments{
\item{.df}{The survey data frame}

\item{question_label}{The label of the question}

\item{group_sep}{The group separator used in the KoBo export}
}
\value{
A list of columns that correspond to each individual possible choice
for \code{question_label}
}
\description{
Get the columns corresponding to a \verb{select multiple} question
}
\examples{
get_question_columns(lby_msna2021, "safety_concerns")
get_question_columns(lby_msna2021, "safety_concerns_boys")
get_question_columns(lby_msna2021, "safety_concerns_girls")
}
