% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{harmonized_municipality_names}
\alias{harmonized_municipality_names}
\title{Harmonized names for Libyan Municipality names}
\format{
An object of class \code{character} of length 8.
}
\source{
\url{reach}
}
\usage{
harmonized_municipality_names
}
\description{
A mapping of Libyan Municipality names and a harmonized
name used by REACH for that Municipality.
}
\keyword{datasets}
