% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{harmonized_neighborhood_names}
\alias{harmonized_neighborhood_names}
\title{Harmonized names for Libyan Neighborhood names}
\format{
An object of class \code{character} of length 2.
}
\source{
\url{reach}
}
\usage{
harmonized_neighborhood_names
}
\description{
A mapping of Libyan Neighborhood names and a harmonized
name used by REACH for that Neighborhood.
}
\keyword{datasets}
