% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiver.R
\name{iview}
\alias{iview}
\title{iview function (interactive view)}
\usage{
iview(X, view = TRUE, save = FALSE, file = NULL)
}
\arguments{
\item{X}{The data, as either a matrix or dataframe.}

\item{view}{Indicate whether the app should be opened in a browser.}

\item{save}{Indicate whether the html file for the app should be saved.}

\item{file}{The filename for the html file if it is to be saved.}
}
\description{
This function passes a matrix or dataframe to a javascript app for interactively viewing data in a table.
}
\examples{
iview(iris)
iview(iris, save=TRUE, file="iris.html")
}
