% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{kobo_assets}
\alias{kobo_assets}
\title{Retrieve a list of KoBoToolbox Assets}
\usage{
kobo_assets(token)
}
\arguments{
\item{token}{KoBo Token}
}
\value{
data frame
}
\description{
Retrieve a list of KoBoToolbox Assets
}
