% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{kobo_download_dataset}
\alias{kobo_download_dataset}
\title{Download survey responses as a dataset}
\usage{
kobo_download_dataset(
  server_url = "https://kobo.humanitarianresponse.info",
  type = "xls",
  lang = "English",
  fields_from_all_versions = TRUE,
  group_sep = "/",
  asset_uid,
  user_name,
  user_password,
  file_path,
  max_retries = 12
)
}
\arguments{
\item{server_url}{URL of the KoBo Server}

\item{type}{Filetype for the export}

\item{lang}{Language of the labels used in the export. Can be 'xls'}

\item{fields_from_all_versions}{Whether to include fields from all versions of the survey}

\item{group_sep}{The character separator for groups}

\item{asset_uid}{The UUID of the survey}

\item{user_name}{The User Name}

\item{user_password}{The Password}

\item{file_path}{The file path to which the dataset is saved}

\item{max_retries}{The number of retries when downloading}
}
\description{
Download survey responses as a dataset
}
