% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{kobo_parse_expression}
\alias{kobo_parse_expression}
\title{Parse a KoBo expression into R.}
\usage{
kobo_parse_expression(statement, as_string = FALSE)
}
\arguments{
\item{statement}{a string representing a KoBo expression}

\item{as_string}{whether to transform the result of parsing into an expression}
}
\value{
an R expression or its string representation
}
\description{
Parse a KoBo expression into R.
}
\examples{
kobo_parse_expression("${hoh}")
kobo_parse_expression("${hoh}", as_string = TRUE)
kobo_parse_expression("selected(${hoh}, 'no')")
kobo_parse_expression("selected(${hoh}, 'no')", as_string = TRUE)
lby_msna2021 \%>\%
  dplyr::filter(!!kobo_parse_expression("selected(${hoh}, 'no')")) \%>\%
  dplyr::distinct(hoh)
}
