% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{kobo_token}
\alias{kobo_token}
\title{Retrieve a KoBoToolbox token for the V2 API}
\usage{
kobo_token(username, password)
}
\arguments{
\item{username}{KoBo username}

\item{password}{KoBo password}
}
\value{
token as a character vector
}
\description{
Retrieve a KoBoToolbox token for the V2 API
}
