% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lby_districts}
\alias{lby_districts}
\title{Libyan Districts}
\format{
An object of class \code{data.frame} with 22 rows and 4 columns.
}
\source{
\url{reach}
}
\usage{
lby_districts
}
\description{
A dataset containing the identifiers as well as
the Arabic and English names of Libyan Districts
}
\details{
\describe{
\item{district_id}{A unique identifier for the District}
\item{district_name_ar}{The Districts' name in Arabic}
\item{district_name_en}{The Districts' name in English}
\item{district_region_id}{The region to which the Districts belongs}
}
}
\keyword{datasets}
