% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lby_msna2021_choices}
\alias{lby_msna2021_choices}
\title{REACH Libya MSNA 2021 Choices (Refugee and Migrant Population)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1351 rows and 7 columns.
}
\source{
\href{https://www.reachresourcecentre.info/}{REACH Resource Centre}
}
\usage{
lby_msna2021_choices
}
\description{
In response to a lack of recent data on the humanitarian situation in Libya,
REACH conducted a multi-sectoral needs assessment (MSNA) from June-July 2021
to provide timely information on the needs and vulnerabilities of refugees
and migrants. REACH designed the data collection and analysis tools, and
consulted with all sectors and working groups active in the Libyan response
to revise indicators.
}
\keyword{datasets}
