% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lby_municipalities}
\alias{lby_municipalities}
\title{Libyan Municipalities}
\format{
An object of class \code{data.frame} with 100 rows and 5 columns.
}
\source{
\url{reach}
}
\usage{
lby_municipalities
}
\description{
A dataset containing the identifiers as well as
the Arabic and English names of Libyan Municipalities
}
\details{
\describe{
\item{municipality_id}{A unique identifier for the Municipality}
\item{municipality_name_ar}{The Municipality's name in Arabic}
\item{municipality_name_en}{The Municipality's name in English}
\item{municipality_district_id}{The district to which the Municipality belongs}
\item{municipality_region_id}{The region to which the Municipality belongs}
}
}
\keyword{datasets}
