% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lby_regions}
\alias{lby_regions}
\title{Libyan Regions}
\format{
An object of class \code{data.frame} with 3 rows and 3 columns.
}
\source{
\url{reach}
}
\usage{
lby_regions
}
\description{
A dataset containing the identifiers as well as
the Arabic and English names of Libyan Regions
}
\details{
\describe{
\item{region_id}{A unique identifier for the Region}
\item{region_name_ar}{The Region's name in Arabic}
\item{region_name_en}{The Region's name in English}
}
}
\keyword{datasets}
