% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiver.R
\name{mplot}
\alias{mplot}
\title{mplot function (multivariate scatterplots)}
\usage{
mplot(X, view = TRUE, save = FALSE, file = NULL)
}
\arguments{
\item{X}{The data, as either a matrix or dataframe.}

\item{view}{Indicate whether the app should be opened in a browser.}

\item{save}{Indicate whether the html file for the app should be saved.}

\item{file}{The filename for the html file if it is to be saved.}
}
\description{
This function passes a matrix or dataframe to a javascript app for interactively viewing scatterplots.
}
\examples{
mplot(iris)
mplot(iris, save=TRUE, file="iris.html")
}
