% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiveR-package.R
\docType{package}
\name{quiveR-package}
\alias{quiveR}
\alias{quiveR-package}
\title{quiveR: A set of tools for interacting with KoBo and analyzing datasets from it}
\description{
A set of tools for interacting with KoBo and analyzing datasets from it.
}
\author{
\strong{Maintainer}: Iyed Ghedamsi \email{contact@iyedg.me}

}
\keyword{internal}
